server/requirements.txt
fastapi
uuid
uvicorn
gunicorn
WebSockets
python-dotenv
redis==3.5.3
setuptools
rejson
